Performance optimizations: DB PRAGMAs, batch inserts, O(n) conflict resolution, -O3 -flto, 369/369 tests

+++ src/core/types/AnalysisResult.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <cstdint>
 #include <map>
 #include <string>
 #include <vector>
@@ -9,7 +10,7 @@ namespace archive::core {
 struct AnalysisResult {
     int total_files = 0;
     int total_folders = 0;
-    int total_size = 0;
+    int64_t total_size = 0;
     std::map<std::string, int> by_extension;
     std::map<std::string, int> by_mime;
     std::map<std::string, int> by_project;
